update to 1.3.0

+++ CppSDK/SDK/KuroImposter_structs.hpp
@@ -26,7 +26,7 @@ enum class EImpPreviewImgRes : uint16
 
 // Enum KuroImposter.ImpMemType
 // NumValues: 0x0004
-enum class EImpMemType : uint32
+enum class EImpMemType : uint8
 {
 	RGBAX2                                   = 0,
 	RGB                                      = 1,
@@ -36,7 +36,7 @@ enum class EImpMemType : uint32
 
 // Enum KuroImposter.ImposterType
 // NumValues: 0x0004
-enum class EImposterType : uint32
+enum class EImposterType : uint8
 {
 	TREE                                     = 0,
 	GENERAL                                  = 1,
